requirements.txt
PyQt5==5.15.9
langchain==0.0.307
langchain-openai==0.1.1
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.0
torch==2.0.1
sentence-transformers==2.2.2
beautifulsoup4==4.12.3
arxiv==2.1.3